Fixed KeyboardKey script and renamed folder
