+++ Area Game Files/Assets/Scripts/Core/Scriptables/Breakable.cs
