+++ Area Game Files/Assets/Scripts/Core/Scriptables/Character.cs
