+++ Area Game Files/Assets/Scripts/Core/Scriptables/KeyboardLayout.cs
@@ -1,0 +1,14 @@
+using UnityEngine;
+
+[CreateAssetMenu(menuName = "Scriptables/KBKeybinding")]
+public class KeyboardLayout : ScriptableObject
+{
+
+    public KeyCode Jump = KeyCode.Space,          
+                   Fire = KeyCode.Mouse0,
+                   Forward = KeyCode.W,
+                   Backwards = KeyCode.S,
+                   StrafeLeft = KeyCode.A,
+                   StrafeRight = KeyCode.D;
+
+}
