+++ Area Game Files/Assets/Scripts/Core/Scriptables/Weapon.cs.meta
